requirements.txt
fastapi
uvicorn
crewai
crewai[tools]
soundfile
speechrecognition
python-dotenv
